Refactor jobs service, add retry/cancel jobs
