+++ crates/libs/common_services/src/api/jobs/mod.rs
@@ -1,2 +1,3 @@
 pub mod interfaces;
 pub mod service;
+pub mod query_helpers;
